[Mod] The size of the image is now specified as float

+++ src/SWF.Core.ImageAccessor/ImageUtil.cs
@@ -136,7 +136,7 @@ namespace SWF.Core.ImageAccessor
             }
         }
 
-        internal static Bitmap Resize(Bitmap srcBmp, int width, int height, OpenCvSharp.InterpolationFlags flag)
+        internal static Bitmap Resize(Bitmap srcBmp, float width, float height, OpenCvSharp.InterpolationFlags flag)
         {
             ArgumentNullException.ThrowIfNull(srcBmp, nameof(srcBmp));
